+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Updates
 
+- Deprecation: Stripe API soft removed.
 - Improvement: Improved the deref operator `*` to be more nimble
 - New Feature: Can now pluck lists of values from collections of nodes and edges.
 - Major Language Feature: Introducing the `yeild` feature. See bible for details